Implement Memoizor library for memoizing functions
This commit adds the implementation of the Memoizor C++ library. The library provides a Memoizor class template that takes a function and returns a new function that caches the results of the original function for given arguments.

The library includes tests to verify the correctness and performance of the Memoizor class. The tests show that the Memoizor class is able to significantly improve the performance of recursive functions by caching intermediate results.

This library can be used to easily memoize any function in C++ and improve its performance for large inputs.